moved types to constants
