--- src/components/UserComponents/types.tsx
@@ -1,5 +1,0 @@
-export const enum SizeType {
-	small = "sm",
-	medium = "md",
-	large = "lg"
-}
